+++ com.unity.netcode.gameobjects/Runtime/Messaging/RpcTargets/NotAuthorityRpcTarget.cs
@@ -29,6 +29,12 @@ namespace Unity.Netcode
                         {
                             continue;
                         }
+
+                        // The CMB-Service holds ID 0 and should not be added to the targets
+                        if (clientId == NetworkManager.ServerClientId && m_NetworkManager.CMBServiceConnection)
+                        {
+                            continue;
+                        }
                         m_GroupSendTarget.Add(clientId);
                     }
                 }
@@ -37,6 +43,12 @@ namespace Unity.Netcode
                     foreach (var clientId in m_NetworkManager.ConnectedClientsIds)
                     {
                         if (clientId == behaviour.OwnerClientId || !networkObject.Observers.Contains(clientId))
+                        {
+                            continue;
+                        }
+
+                        // The CMB-Service holds ID 0 and should not be added to the targets
+                        if (clientId == NetworkManager.ServerClientId && m_NetworkManager.CMBServiceConnection)
                         {
                             continue;
                         }